[Autofill Assistant] Remove WebContents from dependencies
As a result the dependency object stays valid even if the WebContents change which is the case in the UiController. Without this change when the UiController updates WebContents we could not be sure that the activity-related dependencies stayed the same.

Bug: b/173103628
Change-Id: I20f349b92eb8c8833671a552f45312a1df7d4f78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3326941
Auto-Submit: Florian Gauger <fga@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Commit-Queue: Florian Gauger <fga@google.com>
Cr-Commit-Position: refs/heads/main@{#951073}