[Extensions] Fix flakiness for PersistentDynamicContentScripts test
Problem:
• Test fixture sets the testcase name as custom arg inside
SetupOnMainThread function which gets setjust before the test is run and
after browser is initialized.

• The extension reads the custom arg set above to call appropriate
javascript function.

• First test explicitly loads the extension hence it runs fine without
any synchronization issue.

• For second and third test, the extension gets loaded automatically
when the browser object is created for the test.

• The test fails when extension (background script) starts running and
it still can not find the custom args as these two are running on
separate process (synchronization issue)

Solution:
Used ExtensionTestMessageListener to coordinate between C++ test and
background script. The listener needs to be set before browser
object is created otherwise it will not be able to receive messages from
script (for 2nd and 3rd test). Initializing listener object in the setup
function and clearing in teardown function.

Bug: 1269641
Change-Id: I5def44203a4176ca7e179d75def6e98f4ad05933
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329594
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951091}