Background tab adds before group discontinuity.
Previously, a new tab opened in the background from a link would always
add immediately after the last tab associated with the opening page.
However, if one or more intervening tabs are added to a group -
especially one potentially containing unassociated tabs - the new tab
can appear after the group or even inside the group, with the unrelated
tabs.

This change will place the new background tab to the right of the
opening tab, ideally after the final tab associated with the opening
tab, but if there is a group discontinuity, immediately before the
discontinuity instead.

For example:

O = opener
* = associated tab
. = unassociated tab
^ = new tab position
[] = group membership

O * * * * ^ . . .

O * * ^ [* * .] . .

[O * * ^] * * . . .

O * ^ [* *] * . . .

Bug: 1246421
Change-Id: Ife2f1a87f6b9b9fd9517b26dfe175e549fa2df18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3330925
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951138}