Remove unneeded framework work arounds in Android accessibility code
This CL removes some work arounds that have been part of the Android
accessibility code base for some time, but no longer seem to be
needed because the Android framework has patched the bugs. We remove
logic to unfocus and refocus a node during scroll, and a call to
postInvalidate, both of which are not needed at this point. This will
result in a more performant experience because we are generating
less events for AT to process.

AX-Relnotes: N/A
Bug: b/209635489
Change-Id: I2ce65976deee3ff4bbcb00b3ef7965c3c39288ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331370
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#951167}