settings: fix UI issues in security keys credential management dialog
This addresses an issue with the edit page of the "Manage sign-in data"
dialog of the Security Keys settings subpage, where the "username" and
"name" fields had the existing value set as a placeholder. Instead, set
it as the field's value, in consistency with other settings edit
dialogs.

Also fix misalignment of the headers on the overview page.

Fixed: 1275361,1275362
Change-Id: Iffb74c8ecf35c6369cf038b38974c370a87ee75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331516
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#951170}