[AppMenu] Fix the crash in AppMenu#calculateHeightForItems
The bug happens when the |availableScreenSpace| is negative number, and
|heightList| is empty in AppMenu#calculateHeightForItems.
When the Chrome reaches the line 547(previously 525),
  spaceForItems += heightList.get(lastItem++);
the exception java.lang.IndexOutOfBoundsException will be thrown.

|heightList| should never been empty, but it seems happens on some
unpopular devices,

The bug is rarely happened, only 25 crashes since August. But to make
Chrome bulletproof, we can make sure that |availableScreenSpace|
is never smaller than zero, so in this condition, line 542 will not be
true.
  if (availableScreenSpace < spaceForFullItems)

Bug: 1132258
Change-Id: Ic1f282dadab38e62c693939dd8192b9370cbf086
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3326063
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951213}