[ZPS] Start prefetch requests iff no non-prefetch request is in progress
Although not likely, it is possible for the providers to have
an active non-prefetch request when a prefetch request is about
to be started. In such scenarios, starting a prefetch request will
cause the providers to invalidate their active non-prefetch requests
and never get a chance to notify the controller of their status;
thus resulting in the controller to remain in an invalid state.

Bug: 1262373
Change-Id: Iad1bbd7400063d86067854c40beb2faf76422990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331656
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Mohamad Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951225}