AW: remove Api21CompatibilityDelegate and add docs
Now that our minSdkVersion is 23, we should be able to safely remove
Api21CompatibilityDelegate, WebViewDelegateFactory, and some related
code in the initialization path.

Since I did some digging through the code history to verify this removal
is OK, I thought it would also be helpful to document the history of the
initialization path on different OS versions. I added this as class docs
to WebViewChromiumFactoryProvider.

Fixed: 1276584
Test: launch WebView (system_webview_apk) on Android 23 emulator
Test: launch WebView (system_webview_google_apk) on Android 24-28 emulator
Test: launch WebView (trichrome_webview_google_apk) on Android 29+ emulator
Change-Id: I1724d3bcec944f746b796df3f7453e9d066e05dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3315560
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951232}