Implement accessibility for expandable view
Enable 'tab' to access the expandable view.

Add a container for expandable view (event_list_), which is used to
escape from the focusing on the date cell. The same for the
`content_view_`, which is wrapped in the `scroll_view_` (we can see the
scroll view is the container).

screen record:
https://drive.google.com/file/d/1keYIxJtQDxzawhtIM3m_fu5Nql6xDPUu/view?usp=sharing

BUG: 1238927, 1277161
Change-Id: I345a8e016689c2e5c7bb9a0f50403dc97f9fe1a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329081
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951242}