desk_templates: Fix DCHECK failure with chromevox on the name view.
DesksTemplatesItemView observed changes to focus/blur on the name view.
But it is also a button, which is AnimationDelegateView which means it
also observes itself for view changes, which leads to a DCHECK failure.

This CL fixes that by skipping focus/blur changes on itself. A
regression would need to be a browser test and requires some
refactoring with spoken feedback test suite, so fix the crash here
first.

Test: manual
Change-Id: I61c2984b69b0c2c1a4f73a1895a1be76c9052684
Fixed: 1279554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3336505
Reviewed-by: Richard Chui <richui@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951293}