[UIA] Mitigate crash related to TextRangeEndpoints
This CL prevents any further crashes related to
TextRangeEndpoints::OnNodeDeleted and AsValidPosition. It seems like the
range doesn't always move out of deleted nodes before they get deleted,
but it's unclear why at the moment.

To help us understand where this issue is coming from, we've added a
DumpWithoutCrashing() in that function. It will be removed once we
figure out the root cause of the crashes.

Bug: 1278304
Change-Id: I2faef73a26ea55e3706df80bfc3b04e706e436bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331356
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#951336}