Make CodecLogger a template for separation of encoder/decoder statuses
Coming down the pipeline is a change that'll separate out EncoderStatus
from Status (and then eventually DecoderStatus from Status as well).

There really isn't going to be a typesafe way to store the error code
when the codes aren't going to be of the same type, so it makes sense
to instantiate the template on which type is going to be expected.

R=liberato,tguilbert

Bug: 1269577
Change-Id: Ibd20017688c736a5446f67700f002bc38f8f4669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3321659
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951338}