fido/cros: Fix comments in authenticator.cc
The logic to determine if USER_PRESENCE or USER_VERIFICATION request will
be sent to u2fd currently doesn't match its comment. The behavior is
correct but the comment isn't. The correct comment should be "Requests
will be UserPresence if uv is discouraged and power button is enabled,
and be UserVerification otherwise."

Bug: None
Test: None
Change-Id: I17df97ba8c2d5ae25da02108fbce6df81e21428e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329056
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Cr-Commit-Position: refs/heads/main@{#951349}