What's New: Fix session restore case
https://chromium-review.googlesource.com/c/chromium/src/+/3275312
changed the What's New logic for M98 to only open a tab if the
page exists for a given milestone. However, it missed the case of
session restore, where the startup tabs aren't opened by the startup
browser creator and are instead passed to session restore. This CL
updates the session restore logic to match the startup browser
creator logic, so that no What's New tabs are opened unless there is
a page corresponding to the current milestone.

Bug: 1271106
Change-Id: Ibc47ffd21aef42a18a80627d76f7f7fc67ed80cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331237
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951356}