[Zenith] Keep the dialog open when clicking on the device selector
The drop down icon is a little bit small and users might click on the
device selector strip by mistake.

This CL increase the size of the drop down button's ink drop. It also
stops the MousePressed event from bubbling to the MediaItemUI, which
brings the focus back to the tab that is playing the media.

Bug: 1279499
Change-Id: Ib7687023aec1bbf4bdaff760d16fa908aa08ac44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3334986
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Muyao Xu <muyaoxu@google.com>
Cr-Commit-Position: refs/heads/main@{#951372}