[Files SWA]: do not open quick view if nothing is selected
Quick view updates this.entries_ array of selected entries on
changeSingleSelectModeSelection_ event. Independently of this
it reacts to key events. It is possible for quick view to be
triggered when no files are selected, which throws an exception.
This is reported as a crash by crash.corp.google.com. This CL
stops quick view from being displayed if we have no known selected
entries.

Bug: b/210555438
Change-Id: Ic6bd72877763d9eb677b5aae235d338002f01199
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3337393
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951400}