Introduce LayoutNGView, behind a runtime feature flag.
This will be needed by LayoutNG printing.

Because LayoutView() / LayoutNGView() doesn't take an Element parameter,
but rather a Document, we have to change the constructors for all
LayoutObject classes that are inherited by LayoutNGMixin.

There should be quite a few issues to fix before we can enable this, but
for now, just fix the bare minimum to be able to get Chrome to show its
startup page without crashing, and also work with most / some (I tested
a couple, and they worked) sites.

Bug: 1121942
Change-Id: I593b60d3fbceb7e2991347674ae13ac50860b7d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3330276
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951409}