[Launcher pop-in] Add skeleton class of BurnInRanker.
This CL adds a skeleton class for a new BurnInRanker. Follow ups will
add method implementations.

Once implemented, this will be a lightweight ranker at the end of the
ranking stack. It will be responsible for two main tasks:

(1) Enforcing a "burn-in" period of some fixed number of milliseconds,
during which results are collected from providers but are not published
to the front-end. The purpose of this is to prevent the rapid and
jittery updating of the result list at the beginning of display.

(2) Tracking results as belonging to either the pre-burn-in period or
the post-burn-in period. Post-burn-in results will have special
handling (e.g. append-only within a category), with exact details TBD.

Test: builds
Bug: 1279686
Change-Id: I7c170ee3791256a1f3d7f0a20493e5ecf775cddf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3336923
Reviewed-by: Rachel Wong <wrong@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951417}