[MPArch] Migrate proxy and replication state updates to BCS
In many cases, replication state and proxy hosts are updated at the
same time. Excluding frame name and frame policy, most of these
updates occur when there is an existing RenderFrameHost associated
with the FrameTreeNode/RenderFrameHost. As proxy_hosts_ and
replication_state_ exist in BrowsingContextState, the functionality to
update both in the simple cases has been migrated so that it is
stored in BrowsingContextState instead.

Bug: 1270671
Change-Id: I273ea80f53945596309c8313492f0bfb86eea45c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3320249
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Harkiran Bolaria <hbolaria@google.com>
Cr-Commit-Position: refs/heads/main@{#951445}