dlp: Rename DlpContentManager -> DlpContentManagerAsh
As a preparation to introduce DlpContentManager containing
shared functionality in between Ash and Lacros, it's needed
to rename DlpContentManager.
In the future there will be base class with shared functionality
DlpContentManager and classes with platform-specific implementations:
DlpContentManagerAsh and DlpContentManagerLacros.

Tests are not renamed as some of them will be migrated to be enabled
on Lacros too, but this will come in the following CLs after fixing
DLP testing on LaCros.

Bug: 1254326
Test: Trybots
Change-Id: I8d94a58d77e30f0c69048ffd80ba8406e1d58eb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3330363
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951497}