[lacros] fixed popup window size problem.
This is caused by one pixel off when converting between DIP and pixels.
Window geometry is not sent to the server when the visual size update
doesn't match the bounds in pending configure.

The off-pixel comes from the implementation difference in the gfx APIs
ScaleToRoundedRect and ScaleToRoundedSize. It is possible that
gfx::ScaleToRoundedRect(rect).size() !=
gfx::ScaleToRoundedSize(rect.size()).

When setting the window bounds, the former is used, while in
ProcessVisualSizeUpdate the latter is used. I have changed the
conversion to use the same function.

Bug: 1272127
Change-Id: Ibc9b908c62688d61cec0c71b9788ba489dc2d558
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3336510
Auto-Submit: Jun Liu <yjliu@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951502}