Fix the arc::GetWindowTaskOrSessionId crash issue.
From the crash report, we can see the crash increases from 98.0.4729.0,
and guess the potential reason might be related with CL:3295763.

From the call stack, it looks like windows are added to
arc_window_candidates_, but not removed when the window is destroyed.

One guessing is at the very early stage, OnWindowInitialized, the
window might haven't beed added to the browserlist yet, so IgnoreWindow
returns false. But when the window is destroyed, IgnoreWindow returns
true, because the window has been added to the browserlist, and
FindBrowserWithWindow in IgnoreWindow returns true. This might cause
the window is added to arc_window_candidates_, but not removed from
arc_window_candidates_.

This CL removes the checking of IgnoreWindow in OnWindowDestroying,
because if IgnoreWindow returns true in OnWindowInitialized, then
OnWindowDestroying should not be called. Otherwise, we can still go
through the code in OnWindowDestroying.

BUG=1278231

Change-Id: I28d4067c2572fde05b8b9b12fbc3b6e7ff622c32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3337279
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951530}