[Android][Passwords] Remove LoginsResult as JobReturnHandler callback
JobReturnHandler is a class that replies to ongoing backend jobs. We
currently repurpose it to also record metrics for jobs that behave
similar to backend jobs but cannot fail.
That is suboptimal and we introduced an additional callback type
specifically to that (LoginsReply) although this type is already
covered by a more general one (LoginsOrErrorReply).

This CL removes LoginsReply and introduces an adapter. That code
also should only be temporary until the metrics recording method is
independent from the Job-Handling.

Bug: 1229655
Change-Id: I6d9bdb916985233cf3ca0d8b739413d32d824ea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3339312
Reviewed-by: Maria Kazinova <kazinova@google.com>
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951538}