[Android] Send dismissal signal from Blink to browser for DateTimeChooser
On Android, the DateTimeChooser is a browser side Android native popup
window, which is an external DateTimeChooser from Blink's perspective.

When Blink wants to dismiss the DateTimePicker, it will send a
EndChooser() signal. Previously, we didn't wire this signal to browser
process for Android, therefore, when blink wants to dismiss the
popup window, e.g. before navigating away from the current URL, the
popup window will still remain there.

We also guard the Java DateTimeChooser to use the native pointer when
the corresponding native object is gone.

Bug: 1275438
Change-Id: I07f7ca95ed3c39fbb27b1439d3129d91a77a5892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3334126
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951558}