haptics: Add feedback for continuous desks swipe.
There are two cases, edge (different from when trying to start a desk
switch on the edge) and regular. For the regular case, this CL just
piggy backs on the existing visible desk change metric.

For the edge case during a continuous desks swipe, introduce a new
variable which tracks if we are at the edge. If a swipe event causes
us to hit the edge, send a haptic event.

Test: added tests
Bug: 1278944
Change-Id: I1c5bdc5582ab293f84f3feec09d3490069088c43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331518
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951602}