[tab_strip] Preallocate buffer space when generating tab Skia paths
When switching tabs the tab strip redraws all tabs in the top chrome.
When a window has many tabs the time to generate a Skia path for each
tab becomes noticeable. This time comes down to mallocs - each
operation added to the path results in a realloc of the verb and point
buffers inside the path. This cl sizes those buffers large enough at
the start to avoid the successive reallocs.

Bug: 1278509
Change-Id: I4eed9749e3fa1e7195e916cb6abfde1fd36387d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3328883
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jayson Adams <shrike@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951637}